refactor: 重写 `App` 与 `MainPage` 的交互代码，完成基础路由构建
